Add metrics for msbfs
