+++ spark/README.md
@@ -1,0 +1,158 @@
+# Spark Алгоритмы: Boruvka и Multiple Source  Parent BFS
+
+## Общая инструкция
+
+1. **Убедитесь, что установлена Java версии 11:**
+
+   ```bash
+   java -version
+   ```
+
+   При необходимости установите:
+
+   ```bash
+   sudo apt install openjdk-11-jdk
+   ```
+
+2. **Установите sbt (Scala Build Tool):**
+
+   ```bash
+   sudo apt install sbt
+   ```
+
+3. **Запустите sbt:**
+
+   ```bash
+   sbt
+   ```
+
+4. **Внутри sbt вызовите нужную команду в зависимости от требуемого алгоритма.**
+
+
+---
+
+## Multiple Source Parent BFS
+
+### Команда запуска:
+
+```bash
+runMain algos.MultipleSourceParentBFSRun path/to/the/input/file.txt path/to/the/output/file.txt coresNum
+```
+
+* `path/to/the/input/file.txt` — путь к входному файлу с графом
+* `path/to/the/output/file.txt` — путь, по которому будет сохранён результат
+* `coresNum` — количество используемых процессоров (`Int`)
+
+### Формат входного файла:
+
+```
+<s₁> <s₂> ... <sₖ>
+<vertex1>,<vertex2>
+<vertex3>,<vertex4>
+...
+```
+* Первая строка — список стартовых вершин, разделённых пробелом
+* Последующие строки — рёбра графа в формате `<u>,<v>`, представляющие неориентированное ребро между вершинами `u` и `v`.
+
+#### Пример:
+
+```
+1 4
+1,3
+1,2
+2,4
+3,4
+```
+
+### Формат выходного файла:
+
+```
+<source_vertex_1> <parent_for_vertex_1> <parent_for_vertex_2> ... <parent_for_vertex_N>
+...
+```
+
+* Каждая строка соответствует **одной исходной вершине**
+* Количество значений после исходной вершины — равно количеству всех вершин в графе
+* Значения представляют родителя каждой вершины в дереве BFS, построенном от данной исходной вершины
+* **Родитель самой исходной вершины** всегда `-1`
+* Если вершина **недостижима** — указывается `-2`
+
+#### Пример:
+
+```
+1 -1 1 1 2
+4  2 4 4 -1
+```
+
+> Здесь `1` и `4` — исходные вершины. После них следуют родительские вершины для всех остальных (в порядке соответствия номерам вершин графа).
+
+---
+
+## Boruvka (Minimum Spanning Tree)
+
+### Команда запуска:
+
+```bash
+runMain algos.BoruvkaRun path/to/the/input/file.txt path/to/the/output/file.txt coresNum
+```
+
+* `path/to/the/input/file.txt` — путь к входному файлу с графом
+* `path/to/the/output/file.txt` — путь, по которому будет сохранён результат
+* `coresNum` — количество используемых процессоров (`Int`)
+
+### Формат входного файла:
+
+```
+<tag> <src> <dst> <weight>
+...
+```
+
+* `<tag>` — необязательная строка (можно использовать `a` для всех строк)
+* `<src>` — исходная вершина
+* `<dst>` — конечная вершина
+* `<weight>` — вес ребра
+
+#### Пример:
+
+```
+a 0 1 1
+a 1 0 1
+a 0 2 4
+a 2 0 4
+a 1 2 2
+a 2 2 2
+...
+```
+
+### Формат выходного файла:
+
+```
+<src> <dst> <weight>
+```
+
+* Каждая строка — ребро из построенного **минимального остовного дерева (MST)**
+* Вершины и веса представлены в стандартном формате
+
+#### Пример:
+
+```
+0 1 1.0
+1 2 2.0
+```
+---
+
+## Файл с метриками выполнения
+
+Для каждого запуска алгоритма автоматически создаётся файл с метриками времени выполнения. Его имя формируется на основе выходного пути, например:
+
+```
+output path: src/main/scala/algos/example.txt
+=> metrics path: src/main/resources/metrics/example_msbfs_algorithm_timing.txt
+```
+
+### Формат метрик:
+
+```
+ComputationTime(ms): <число>
+CommunicationTime(ms): <число>
+```
